backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.0.1
psycopg2-binary==2.9.10
PyJWT==2.8.0
pydantic==2.11.7
email-validator==2.1.1
Flask-Cors==4.0.0
